+++ saleor/graphql/account/deprecated/types.py
@@ -5,7 +5,6 @@ from ....app.models import App, AppToken
 from ....core.permissions import AppPermission
 from ...core.connection import CountableDjangoObjectType
 from ...core.types import FilterInputObjectType, Permission
-from ...meta.deprecated.resolvers import resolve_meta, resolve_private_meta
 from ...meta.types import ObjectWithMetadata
 from ...utils import format_permissions_for_display
 from .filters import ServiceAccountFilter
@@ -74,14 +73,6 @@ class ServiceAccount(CountableDjangoObjectType):
     @staticmethod
     def resolve_tokens(root: App, _info, **_kwargs):
         return root.tokens.all()  # type: ignore
-
-    @staticmethod
-    def resolve_meta(root: App, info):
-        return resolve_meta(root, info)
-
-    @staticmethod
-    def resolve_private_meta(root: App, _info):
-        return resolve_private_meta(root, _info)
 
     @staticmethod
     def __resolve_reference(root, _info, **_kwargs):