+++ saleor/graphql/order/schema.py
@@ -21,10 +21,6 @@ from .mutations.draft_orders import (
 )
 from .mutations.fulfillments import (
     FulfillmentCancel,
-    FulfillmentClearMeta,
-    FulfillmentClearPrivateMeta,
-    FulfillmentUpdateMeta,
-    FulfillmentUpdatePrivateMeta,
     FulfillmentUpdateTracking,
     OrderFulfill,
 )
@@ -32,13 +28,9 @@ from .mutations.orders import (
     OrderAddNote,
     OrderCancel,
     OrderCapture,
-    OrderClearMeta,
-    OrderClearPrivateMeta,
     OrderMarkAsPaid,
     OrderRefund,
     OrderUpdate,
-    OrderUpdateMeta,
-    OrderUpdatePrivateMeta,
     OrderUpdateShipping,
     OrderVoid,
 )
@@ -159,61 +151,12 @@ class OrderMutations(graphene.ObjectType):
     order_add_note = OrderAddNote.Field()
     order_cancel = OrderCancel.Field()
     order_capture = OrderCapture.Field()
-    order_clear_private_meta = OrderClearPrivateMeta.Field(
-        deprecation_reason=(
-            "Use the `deletePrivateMetadata` mutation instead. This field will be "
-            "removed after 2020-07-31."
-        )
-    )
-    order_clear_meta = OrderClearMeta.Field(
-        deprecation_reason=(
-            "Use the `deleteMetadata` mutation instead. This field will be removed "
-            "after 2020-07-31."
-        )
-    )
     order_fulfill = OrderFulfill.Field()
     order_fulfillment_cancel = FulfillmentCancel.Field()
     order_fulfillment_update_tracking = FulfillmentUpdateTracking.Field()
-    order_fulfillment_clear_meta = FulfillmentClearMeta.Field(
-        deprecation_reason=(
-            "Use the `deleteMetadata` mutation instead. This field will be removed "
-            "after 2020-07-31."
-        )
-    )
-    order_fulfillment_clear_private_meta = FulfillmentClearPrivateMeta.Field(
-        deprecation_reason=(
-            "Use the `deletePrivateMetadata` mutation instead. This field will be "
-            "removed after 2020-07-31."
-        )
-    )
-    order_fulfillment_update_meta = FulfillmentUpdateMeta.Field(
-        deprecation_reason=(
-            "Use the `updateMetadata` mutation instead. This field will be removed "
-            "after 2020-07-31."
-        )
-    )
-    order_fulfillment_update_private_meta = FulfillmentUpdatePrivateMeta.Field(
-        deprecation_reason=(
-            "Use the `updatePrivateMetadata` mutation instead. This field will be "
-            "removed after 2020-07-31."
-        )
-    )
     order_mark_as_paid = OrderMarkAsPaid.Field()
     order_refund = OrderRefund.Field()
     order_update = OrderUpdate.Field()
-    order_update_meta = OrderUpdateMeta.Field(
-        deprecation_reason=(
-            "Use the `updateMetadata` mutation instead. This field will be removed "
-            "after 2020-07-31."
-        )
-    )
-    order_update_private_meta = OrderUpdatePrivateMeta.Field(
-        deprecation_reason=(
-            "Use the `updatePrivateMetadata` mutation instead. This field will be "
-            "removed after 2020-07-31."
-        )
-    )
     order_update_shipping = OrderUpdateShipping.Field()
     order_void = OrderVoid.Field()
-
     order_bulk_cancel = OrderBulkCancel.Field()