+++ saleor/urls.py
@@ -27,6 +27,7 @@ urlpatterns = [
 
 if settings.DEBUG:
     import warnings
+
     from .core import views
 
     try: