+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable, unreleased changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Add metadata to shipping zones and shipping methods - #6340 by @maarcingebala
+- Drop deprecated meta mutations - #6422 by @maarcingebala
+
 ## 2.11.0
 
 ### Features